+++ shotmanager/utils/utils.py
@@ -834,3 +834,14 @@ def segment_is_in_range(segment_start, segment_end, range_start, range_end, part
 def clamp(value, minimum, maximum):
     return min(max(value, minimum), maximum)
 
+
+def remap(value, old_min, old_max, new_min, new_max):
+    value = clamp(value, old_min, old_max)
+    old_range = old_max - old_min
+    if old_range == 0:
+        new_value = new_min
+    else:
+        new_range = new_max - new_min
+        new_value = (((value - old_min) * new_range) / old_range) + new_min
+
+    return new_value